Lodha Manyata Tech Park Bangalore
Renowned real estate developers Lodha Group has launched its mega residential project named Lodha Manyata Tech Park in the buzzing area of Bangalore. This residential development in Thanisandra Main Road has a lot to offer; One of them is its 85% open space with a beautifully landscaped garden area. Amidst the hustle and bustle of the city, this place offers a serene environment with a lush green neighborhood which helps the residents to rejuvenate in the new environment. Leveraging cutting-edge technology, innovative interiors, and strong architectural prowess, this gated township in the heart of Bangalore is one of the best locales that offers opulent luxury at an affordable price.

 

Lodha Manyata Tech Park in Bangalore, 2BHK, 3BHK and 4BHK apartments ranging from various sq. ft. are innovatively designed to redefine the living standards of the residents. The spacious residences are well expressed with vitrified floors and heavy powder-coated aluminum sliding windows that give a panoramic view of the city. Textured emulsion paint has been used for the living room and bedroom while ceramic tiles, beautiful counters , and modular cabinetry adorn the galley. In addition, the bathrooms have used superior excellence fixtures and fittings, and a laminate wooden floor with an uninterrupted view from the gallery is soothing to the eyes.

 

 Apart from the opulent interiors, the first visual impression of this housing project Lodha Manyata Tech Park Bangalore begins from the exclusive entrance lobby of this gated development. The project comprises a vast patch of lush green landscape with a beautiful garden, water bodies and seating arrangements.

 

For entertainment, this residential development has a child's play area, indoor game room and outdoor game court, and a luxurious clubhouse with a swimming pool and etc. In addition, there is a gym and meditation deck for fitness as well as a multipurpose hall for hosting. events here. The project provides a multi-layered security system for the safety of the residents. Other amenities include a car parking area, 24-hour electricity and water supply, lift and fire fighting equipment.

 

This property is located at a strategic position in Thanisandra Main Road, just minutes away from developments. The location enjoys proximity to various important hubs and commercial areas of Bangalore. Apart from this it also has access to various colleges, culinary destinations, clinics, supermarkets and multiplexes.

For More Details 
Visit us -  https://bit.ly/3XVotIH